create scorer module
