+++ lib/search.js
@@ -1,10 +1,9 @@
 var util = require('./util');
+var scorer = require('./scorer');
 
 var SearchController = {};
 SearchController.getSuggestions = function(cities, options, callback) {
   var startTime = new Date().getTime();
-  //console.log('SearchController.getSuggestions with query: ', options);
-
   var suggestions = [];
   var err= '';
 
@@ -14,12 +13,12 @@ SearchController.getSuggestions = function(cities, options, callback) {
   options.latitude = parseFloat(options.latitude) || null;
   options.longitude = parseFloat(options.longitude) || null;
 
-  //Loop on the whole array of cities to cal
+  //Loop on the whole array of cities to calculate score
   cities.forEach(function(city) {
     // check only city with population
     if (city.population >= 5000) {
       var geoData  = { latitude: options.latitude, longitude: options.longitude };
-      var score = util.getScore(city,queryString,geoData);
+      var score = scorer.getScore(city,queryString,geoData);
 
       // if score superior to 0 add result to suggestions
       if (score > 0 ) {
